TODO-49 delete user check

+++ main/functions/lists.php
@@ -95,7 +95,7 @@ function deleteList(){
     $obj = json_decode($_POST["json"], false);
     $list_id = $obj->listID;
 
-    $stmt = $dbh->prepare('DELETE FROM listy WHERE id = :list_id');
+    $stmt = $dbh->prepare('DELETE FROM listy WHERE id = :list_id AND user_id = :user_id');
 
     $stmt->bindParam(':list_id', $list_id);
